Refactor: Improve error handling and context in LocalCodeExecutor and console output
